+++ model.go
@@ -24,6 +24,8 @@ var (
 	TagWeightsJSON        = config.TagWeightsJSON
 	ExtraTagWeightsString = config.ExtraTagWeightsString
 	ExtraTagWeightsJSON   = config.ExtraTagWeightsJSON
+	ExcludeTagsString     = config.ExcludeTagsString
+	AllTagWeights         = config.AllTagWeights
 	AdjustScores          = config.AdjustScores
 	Extensions            = config.Extensions